@@ -1,6 +1,6 @@
 {
   "name": "mkdocs-material",
-  "version": "7.1.9+insiders-2.9.2",
+  "version": "7.2.0+insiders-2.11.1",
   "description": "A Material Design theme for MkDocs",
   "keywords": [
     "mkdocs",
@@ -39,7 +39,7 @@
     "fuzzaldrin-plus": "^0.6.0",
     "lunr": "^2.3.9",
     "lunr-languages": "^1.8.0",
-    "rxjs": "^7.1.0"
+    "rxjs": "^7.2.0"
   },
   "devDependencies": {
     "@fortawesome/fontawesome-free": "^5.15.3",
@@ -47,22 +47,22 @@
     "@primer/octicons": "^14.2.2",
     "@types/clipboard": "^2.0.7",
     "@types/escape-html": "1.0.1",
-    "@types/fuzzaldrin-plus": "^0.6.1",
-    "@types/html-minifier": "^4.0.0",
-    "@types/lunr": "^2.3.3",
-    "@types/node": "^15.12.5",
-    "@types/resize-observer-browser": "^0.1.5",
-    "@types/sass": "^1.16.0",
-    "@typescript-eslint/eslint-plugin": "^4.28.1",
-    "@typescript-eslint/parser": "^4.28.1",
-    "autoprefixer": "^10.2.6",
+    "@types/fuzzaldrin-plus": "^0.6.2",
+    "@types/html-minifier": "^4.0.1",
+    "@types/lunr": "^2.3.4",
+    "@types/node": "^16.3.3",
+    "@types/resize-observer-browser": "^0.1.6",
+    "@types/sass": "^1.16.1",
+    "@typescript-eslint/eslint-plugin": "^4.28.3",
+    "@typescript-eslint/parser": "^4.28.3",
+    "autoprefixer": "^10.3.1",
     "chokidar": "^3.5.2",
     "cssnano": "^5.0.6",
-    "esbuild": "^0.12.12",
-    "eslint": "^7.29.0",
+    "esbuild": "^0.12.15",
+    "eslint": "^7.31.0",
     "eslint-plugin-eslint-comments": "^3.2.0",
     "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jsdoc": "^35.4.1",
+    "eslint-plugin-jsdoc": "^35.4.5",
     "eslint-plugin-no-null": "^1.0.2",
     "github-types": "^1.0.0",
     "gitlab": "^14.2.2",
@@ -71,9 +71,9 @@
     "material-shadows": "^3.0.1",
     "postcss": "^8.3.5",
     "postcss-inline-svg": "^5.0.0",
-    "preact": "^10.5.13",
+    "preact": "^10.5.14",
     "rimraf": "^3.0.2",
-    "sass": "^1.35.1",
+    "sass": "^1.35.2",
     "stylelint": "^13.13.1",
     "stylelint-config-rational-order": "^0.1.2",
     "stylelint-config-recommended": "^5.0.0",
@@ -81,8 +81,8 @@
     "stylelint-scss": "^3.19.0",
     "svgo": "^2.3.1",
     "tiny-glob": "^0.2.9",
-    "ts-node": "^10.0.0",
-    "typescript": "^4.3.4"
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5"
   },
   "engines": {
     "node": ">= 14"
